requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
supabase>=1.0.3
python-dotenv>=1.0.0
gunicorn>=21.0.0
